+++ src/main/java/com/stockmarket/StockMarketSimulator/repositories/InvestorRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.stockmarket.StockMarketSimulator.model.Investor;
 
-public interface InvestorRepository extends JpaRepository<Investor, Long>{
+public interface InvestorRepository extends JpaRepository<Investor, Integer>{
 	
-	List<Investor> findInvestorID(Long investorId);
+	List<Investor> findInvestorID(Integer investorId);
 
 }
